move weakarraylist to util instead of main package
